refactor(disagg): hoist duplicated _handle_staging_req into a mixin
`_handle_staging_req` was byte-identical, 33 lines, in both
`MooncakeKVManager` and `NixlKVManager` -- `diff` reported zero differences.
It was a straight copy-paste, and tellingly it already delegated to
`common/staging_handler.handle_staging_req`, so the shared helper existed and
only the wrapper had been duplicated.

Moves it to a `StagingManagerMixin` in `common/staging_handler.py`, beside
the `handle_staging_req` it wraps, and mixes it into the two managers that
receive STAGING_REQ. The method body is unchanged apart from dropping its
local `from ...staging_handler import handle_staging_req`, which is now
redundant because the mixin lives in that module.

Chose a mixin over adding the method to `CommonKVManager`: Mori also extends
`CommonKVManager` but has no staging support (no `_staging_ctx`), so putting
it on the shared base would hand Mori a method it cannot honor. The mixin
docstring states the attributes a concrete manager must provide.

`AscendKVManager` extends `MooncakeKVManager` and so inherits the method
through the mixin, unchanged.

One behavior delta worth noting: the "unregistered room" warning is now
emitted on the `common.staging_handler` logger rather than the per-backend
`mooncake.conn` / `nixl.conn` logger, so that log line's module attribution
changes. Nothing filters on it.

No functional change.
